test: migrate test fixture/helper
